backend/event_logic/requirements.txt
fastapi==0.115.4
uvicorn==0.32.0
pydantic==2.9.2
SQLAlchemy==2.0.36
PyJWT==2.9.0
python-dotenv==1.0.1
aiosqlite==0.20.0
pydantic_settings==2.6.0
loguru==0.7.3
asyncpg==0.30.0